% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{age}
\alias{age}
\title{Calculate Age Fast}
\usage{
age(from, to)
}
\arguments{
\item{from}{character string (or vector), like '1994-10-17'}

\item{to}{character string (or vector), like '2019-06-22'}
}
\description{
Calculate Age Fast
}
