% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{calc_ci}
\alias{calc_ci}
\title{Calculate Confidential Inteval}
\usage{
calc_ci(sample, interval = 95)
}
\arguments{
\item{sample}{事後サンプルベクトル}

\item{interval}{信頼区間の\%}
}
\description{
Calculate Confidential Inteval
}
