% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inverse_map}
\alias{inverse_map}
\title{Apply Multiple Functions to ONE Vector}
\usage{
inverse_map(x, ...)
}
\arguments{
\item{x}{A Vector}

\item{...}{Function Names}
}
\value{
list
}
\description{
Apply Multiple Functions to ONE Vector
}
\examples{
inverse_map(c(20, 100, 30), min, max)
}
