% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs.R
\name{is_installed}
\alias{is_installed}
\title{Whether a Pakckage is Installed}
\usage{
is_installed(pkg)
}
\arguments{
\item{pkg}{string of the pakcage name to find}
}
\value{
Logical
}
\description{
Whether a Pakckage is Installed
}
