% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{kansuji2int}
\alias{kansuji2int}
\title{Kansuji to Int
cf. https://gist.github.com/kos59125/bec05eab7c33072dc008}
\usage{
kansuji2int(x)
}
\arguments{
\item{x}{A Character}
}
\description{
Kansuji to Int
cf. https://gist.github.com/kos59125/bec05eab7c33072dc008
}
