% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{mecab_wakati}
\alias{mecab_wakati}
\title{MeCab Wakati}
\usage{
mecab_wakati(str, tagger_param = list(l = 2, d = NULL), extract_pattern = NULL)
}
\arguments{
\item{str}{string vector}

\item{tagger_param}{parameters of MeCab Tagger}

\item{extract_pattern}{POS regex}
}
\description{
MeCab Wakati
}
\examples{
\dontrun{
meacb_wakati(string = "私は大きい大学へ行く",
  tagger_param = list(d = "/usr/local/lib/mecab/dic/mecab-ipadic-neologd"),
  extract_pattern = "名詞|形容詞")
}
}
