% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{mecab_wakati_impl}
\alias{mecab_wakati_impl}
\title{MeCab Wakati Impl}
\usage{
mecab_wakati_impl(
  str,
  tagger_param = list(l = 2, d = NULL),
  extract_pattern = NULL
)
}
\arguments{
\item{str}{string}

\item{tagger_param}{parameters of MeCab Tagger}

\item{extract_pattern}{POS regex}
}
\description{
MeCab Wakati Impl
}
