% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs.R
\name{min_r_version}
\alias{min_r_version}
\title{Find minimum R version required for package}
\source{
Based on ideas from
  http://stackoverflow.com/questions/38686427/determine-minimum-r-version-for-all-package-dependencies
}
\usage{
min_r_version(pkg)
}
\arguments{
\item{pkg}{string with name of package to check}
}
\description{
Recursively search dependencies for R version, and find the highest stated R
version requirement.
}
\examples{
base <- c(
  "base", "compiler", "datasets", "grDevices", "graphics",
  "grid", "methods", "parallel", "profile", "splines", "stats",
  "stats4", "tcltk", "tools", "translations"
)
\dontrun{
base_reqs <- lapply(base, min_r_version)
contrib <- c(
  "KernSmooth", "MASS", "Matrix", "boot",
  "class", "cluster", "codetools", "foreign", "lattice",
  "mgcv", "nlme", "nnet", "rpart", "spatial", "survival"
)
contrib_reqs <- lapply(contrib, min_r_version)
min_r_version("icd")
}
}
