% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{plot_roc}
\alias{plot_roc}
\title{Calculate AUC and Plot ROC Curve}
\usage{
plot_roc(actual, score)
}
\arguments{
\item{actual}{真値のベクトル}

\item{score}{予測値のベクトル}
}
\description{
Calculate AUC and Plot ROC Curve
}
