% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymattuR.R
\docType{package}
\name{ymattuR-package}
\alias{ymattuR}
\alias{ymattuR-package}
\title{ymattuR: Utilities for ymattu}
\description{
Tools library which @y__mattu often uses. Almost all functions are
    designed for @y__mattu.
}
\author{
\strong{Maintainer}: Yuya Matsumura \email{mattu.yuya@gmail.com}

}
\keyword{internal}
