% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymattuR.R
\docType{package}
\name{ymattuR}
\alias{ymattuR}
\alias{ymattuR-package}
\title{\code{ymattuR} package}
\description{
Utilities for y__mattu
}
\details{
See the README on
\href{https://github.com/ymattu/ymattuR#readme}{GitHub}
}
\author{
\strong{Maintainer}: Yuya Matsumura \email{mattu.yuya@gmail.com}

}
\keyword{internal}
